streamline voro file and class names
