+++ SKIRT/voro/voro_compute.cc
@@ -4,10 +4,10 @@
 // Email    : chr@alum.mit.edu
 // Date     : August 30th 2011
 
-/* \file v_compute.cc
- * \brief Function implementations for the voro_compute class. */
+/* \file voro_compute.cc
+ * \brief Function implementations for the Voronoi compute class. */
 
-#include "v_compute.hh"
+#include "voro_compute.hh"
 
 namespace voro {
 
@@ -15,7 +15,7 @@ namespace voro {
  * sets up the mask and queue used for Voronoi computations.
  * \param[in] con_ a reference to the container class to use.
  * \param[in] (hx_,hy_,hz_) the size of the mask to use. */
-voro_compute::voro_compute(container &con_,int hx_,int hy_,int hz_) :
+compute::compute(container &con_,int hx_,int hy_,int hz_) :
     con(con_), boxx(con_.boxx), boxy(con_.boxy), boxz(con_.boxz),
     xsp(con_.xsp), ysp(con_.ysp), zsp(con_.zsp),
     hx(hx_), hy(hy_), hz(hz_), hxy(hx_*hy_), hxyz(hxy*hz_), ps(con_.ps),
@@ -38,7 +38,7 @@ voro_compute::voro_compute(container &con_,int hx_,int hy_,int hz_) :
  *		    vector is within.
  * \param[in,out] mrs the current minimum distance, that may be updated if a
  * 		      closer particle is found. */
-void voro_compute::scan_all(int ijk,double x,double y,double z,int di,int dj,int dk,particle_record &w,double &mrs) {
+void compute::scan_all(int ijk,double x,double y,double z,int di,int dj,int dk,particle_record &w,double &mrs) {
     double x1,y1,z1,rs;bool in_block=false;
     for(int l=0;l<co[ijk];l++) {
         x1=p[ijk][ps*l]-x;
@@ -56,7 +56,7 @@ void voro_compute::scan_all(int ijk,double x,double y,double z,int di,int dj,int
  * will definitely have enough memory to add six entries at the end.
  * \param[in] (ei,ej,ek) the block to consider.
  * \param[in,out] qu_e a pointer to the end of the queue. */
-void voro_compute::add_to_mask(int ei,int ej,int ek,int *&qu_e) {
+void compute::add_to_mask(int ei,int ej,int ek,int *&qu_e) {
     unsigned int *mijk=mask+ei+hx*(ej+hy*ek);
     if(ek>0) if(*(mijk-hxy)!=mv) {if(qu_e==qu_l) qu_e=qu;*(mijk-hxy)=mv;*(qu_e++)=ei;*(qu_e++)=ej;*(qu_e++)=ek-1;}
     if(ej>0) if(*(mijk-hx)!=mv) {if(qu_e==qu_l) qu_e=qu;*(mijk-hx)=mv;*(qu_e++)=ei;*(qu_e++)=ej-1;*(qu_e++)=ek;}
@@ -71,7 +71,7 @@ void voro_compute::add_to_mask(int ei,int ej,int ek,int *&qu_e) {
  * \param (mijk) ?
  * \param[in] (ei,ej,ek) the block to consider.
  * \param[in,out] qu_e a pointer to the end of the queue. */
-void voro_compute::scan_bits_mask_add(unsigned int q,unsigned int *mijk,int ei,int ej,int ek,int *&qu_e) {
+void compute::scan_bits_mask_add(unsigned int q,unsigned int *mijk,int ei,int ej,int ek,int *&qu_e) {
     const unsigned int b1=1<<21,b2=1<<22,b3=1<<24,b4=1<<25,b5=1<<27,b6=1<<28;
     if((q&b2)==b2) {
         if(ei>0) {*(mijk-1)=mv;*(qu_e++)=ei-1;*(qu_e++)=ej;*(qu_e++)=ek;}
@@ -110,7 +110,7 @@ void voro_compute::scan_bits_mask_add(unsigned int q,unsigned int *mijk,int ei,i
  *                       in relative to the container data structure.
  * \return False if the Voronoi cell was completely removed during the
  *         computation and has zero volume, true otherwise. */
-bool voro_compute::compute_cell(voronoicell &c,int ijk,int s,int ci,int cj,int ck) {
+bool compute::compute_cell(cell &c,int ijk,int s,int ci,int cj,int ck) {
     static const int count_list[8]={7,11,15,19,26,35,45,59},*count_e=count_list+8;
     double x,y,z,x1,y1,z1,qx=0,qy=0,qz=0;
     double xlo,ylo,zlo,xhi,yhi,zhi,x2,y2,z2,rs;
@@ -451,7 +451,7 @@ bool voro_compute::compute_cell(voronoicell &c,int ijk,int s,int ci,int cj,int c
  * \param[in] (xh,yh,zh) the relative coordinates of the corner of the block
  *                       furthest away from the cell center.
  * \return False if the block may intersect, true if does not. */
-bool voro_compute::corner_test(voronoicell &c,double xl,double yl,double zl,double xh,double yh,double zh) {
+bool compute::corner_test(cell &c,double xl,double yl,double zl,double xh,double yh,double zh) {
     con.r_prime(xl*xl+yl*yl+zl*zl);
     if(c.plane_intersects_guess(xh,yl,zl,con.r_cutoff(xl*xh+yl*yl+zl*zl))) return false;
     if(c.plane_intersects(xh,yh,zl,con.r_cutoff(xl*xh+yl*yh+zl*zl))) return false;
@@ -474,7 +474,7 @@ bool voro_compute::corner_test(voronoicell &c,double xl,double yl,double zl,doub
  * \param[in] (yh,zh) the relative y and z coordinates of the corner of the
  *                    block furthest away from the cell center.
  * \return False if the block may intersect, true if does not. */
-bool voro_compute::edge_x_test(voronoicell &c,double x0,double yl,double zl,double x1,double yh,double zh) {
+bool compute::edge_x_test(cell &c,double x0,double yl,double zl,double x1,double yh,double zh) {
     con.r_prime(yl*yl+zl*zl);
     if(c.plane_intersects_guess(x0,yl,zh,con.r_cutoff(yl*yl+zl*zh))) return false;
     if(c.plane_intersects(x1,yl,zh,con.r_cutoff(yl*yl+zl*zh))) return false;
@@ -497,7 +497,7 @@ bool voro_compute::edge_x_test(voronoicell &c,double x0,double yl,double zl,doub
  * \param[in] (xh,zh) the relative x and z coordinates of the corner of the
  *                    block furthest away from the cell center.
  * \return False if the block may intersect, true if does not. */
-bool voro_compute::edge_y_test(voronoicell &c,double xl,double y0,double zl,double xh,double y1,double zh) {
+bool compute::edge_y_test(cell &c,double xl,double y0,double zl,double xh,double y1,double zh) {
     con.r_prime(xl*xl+zl*zl);
     if(c.plane_intersects_guess(xl,y0,zh,con.r_cutoff(xl*xl+zl*zh))) return false;
     if(c.plane_intersects(xl,y1,zh,con.r_cutoff(xl*xl+zl*zh))) return false;
@@ -519,7 +519,7 @@ bool voro_compute::edge_y_test(voronoicell &c,double xl,double y0,double zl,doub
  * \param[in] (xh,yh) the relative x and y coordinates of the corner of the
  *                    block furthest away from the cell center.
  * \return False if the block may intersect, true if does not. */
-bool voro_compute::edge_z_test(voronoicell &c,double xl,double yl,double z0,double xh,double yh,double z1) {
+bool compute::edge_z_test(cell &c,double xl,double yl,double z0,double xh,double yh,double z1) {
     con.r_prime(xl*xl+yl*yl);
     if(c.plane_intersects_guess(xl,yh,z0,con.r_cutoff(xl*xl+yl*yh))) return false;
     if(c.plane_intersects(xl,yh,z1,con.r_cutoff(xl*xl+yl*yh))) return false;
@@ -540,7 +540,7 @@ bool voro_compute::edge_z_test(voronoicell &c,double xl,double yl,double z0,doub
  * \param[in] (z0,z1) the minimum and maximum relative z coordinates of the
  *                    block.
  * \return False if the block may intersect, true if does not. */
-bool voro_compute::face_x_test(voronoicell &c,double xl,double y0,double z0,double y1,double z1) {
+bool compute::face_x_test(cell &c,double xl,double y0,double z0,double y1,double z1) {
     con.r_prime(xl*xl);
     if(c.plane_intersects_guess(xl,y0,z0,con.r_cutoff(xl*xl))) return false;
     if(c.plane_intersects(xl,y0,z1,con.r_cutoff(xl*xl))) return false;
@@ -559,7 +559,7 @@ bool voro_compute::face_x_test(voronoicell &c,double xl,double y0,double z0,doub
  * \param[in] (z0,z1) the minimum and maximum relative z coordinates of the
  *                    block.
  * \return False if the block may intersect, true if does not. */
-bool voro_compute::face_y_test(voronoicell &c,double x0,double yl,double z0,double x1,double z1) {
+bool compute::face_y_test(cell &c,double x0,double yl,double z0,double x1,double z1) {
     con.r_prime(yl*yl);
     if(c.plane_intersects_guess(x0,yl,z0,con.r_cutoff(yl*yl))) return false;
     if(c.plane_intersects(x0,yl,z1,con.r_cutoff(yl*yl))) return false;
@@ -578,7 +578,7 @@ bool voro_compute::face_y_test(voronoicell &c,double x0,double yl,double z0,doub
  * \param[in] (y0,y1) the minimum and maximum relative y coordinates of the
  *                    block.
  * \return False if the block may intersect, true if does not. */
-bool voro_compute::face_z_test(voronoicell &c,double x0,double y0,double zl,double x1,double y1) {
+bool compute::face_z_test(cell &c,double x0,double y0,double zl,double x1,double y1) {
     con.r_prime(zl*zl);
     if(c.plane_intersects_guess(x0,y0,zl,con.r_cutoff(zl*zl))) return false;
     if(c.plane_intersects(x0,y1,zl,con.r_cutoff(zl*zl))) return false;
@@ -601,7 +601,7 @@ bool voro_compute::face_z_test(voronoicell &c,double x0,double y0,double zl,doub
  * \param[in] mrs the distance to be tested.
  * \return True if the region is further away than mrs, false if the region in
  *         within mrs. */
-bool voro_compute::compute_min_max_radius(int di,int dj,int dk,double fx,double fy,double fz,double gxs,double gys,double gzs,double &crs,double mrs) {
+bool compute::compute_min_max_radius(int di,int dj,int dk,double fx,double fy,double fz,double gxs,double gys,double gzs,double &crs,double mrs) {
     double xlo,ylo,zlo;
     if(di>0) {
         xlo=di*boxx-fx;
@@ -750,7 +750,7 @@ bool voro_compute::compute_min_max_radius(int di,int dj,int dk,double fx,double 
     return false;
 }
 
-bool voro_compute::compute_min_radius(int di,int dj,int dk,double fx,double fy,double fz,double mrs) {
+bool compute::compute_min_radius(int di,int dj,int dk,double fx,double fy,double fz,double mrs) {
     double t,crs;
 
     if(di>0) {t=di*boxx-fx;crs=t*t;}
@@ -769,7 +769,7 @@ bool voro_compute::compute_min_radius(int di,int dj,int dk,double fx,double fy,d
 /** Adds memory to the queue.
  * \param[in,out] qu_s a reference to the queue start pointer.
  * \param[in,out] qu_e a reference to the queue end pointer. */
-void voro_compute::add_list_memory(int*& qu_s,int*& qu_e) {
+void compute::add_list_memory(int*& qu_s,int*& qu_e) {
     qu_size<<=1;
     int *qu_n=new int[qu_size],*qu_c=qu_n;
     if(qu_s<=qu_e) {
