+++ SKIRT/voro/voro_config.hh
@@ -4,11 +4,11 @@
 // Email    : chr@alum.mit.edu
 // Date     : August 30th 2011
 
-/* \file config.hh
+/* \file voro_config.hh
  * \brief Master configuration file for setting various compile-time options. */
 
-#ifndef VOROPP_CONFIG_HH
-#define VOROPP_CONFIG_HH
+#ifndef VORO_CONFIG_HH
+#define VORO_CONFIG_HH
 
 namespace voro {
 
@@ -87,13 +87,28 @@ const double optimal_particles=5.6;
 
 /** Voro++ returns this status code if there is a memory allocation error, if
  * one of the safe memory limits is exceeded. */
-#define VOROPP_MEMORY_ERROR 2
+const int VOROPP_MEMORY_ERROR=2;
 
 /** Voro++ returns this status code if there is any type of internal error, if
  * it detects that representation of the Voronoi cell is inconsistent. This
  * status code will generally indicate a bug, and the developer should be
  * contacted. */
-#define VOROPP_INTERNAL_ERROR 3
+const int VOROPP_INTERNAL_ERROR=3;
+
+// These constants are automatically generated by worklist_gen.pl and are not
+// intended to be edited by hand.
+
+/** Each region is divided into a grid of subregions, and a worklist is
+# constructed for each. This parameter sets is set to half the number of
+# subregions that the block is divided into. */
+const int wl_hgrid=4;
+/** The number of subregions that a block is subdivided into, which is twice
+the value of hgrid. */
+const int wl_fgrid=8;
+/** The total number of worklists, set to the cube of hgrid. */
+const int wl_hgridcu=64;
+/** The number of elements in each worklist. */
+const int wl_seq_length=64;
 
 }
