+++ SKIRT/voro/voro_container.cc
@@ -4,13 +4,15 @@
 // Email    : chr@alum.mit.edu
 // Date     : August 30th 2011
 
-/* \file v_base_wl.hh
- * \brief The table of block worklists that are used during the cell
- * computation, which is part of the voro_base class.
- *
- * This file is automatically generated by worklist_gen.pl and it is not
- * intended to be edited by hand. */
+/* \file voro_container.cc
+ * \brief Function implementations for the Voronoi container class. */
 
+#include "voro_container.hh"
+
+namespace voro {
+
+// These data are automatically generated by worklist_gen.pl and are not
+// intended to be edited by hand.
 const unsigned int voro_base::wl[wl_seq_length*wl_hgridcu]={
     7,0x10203f,0x101fc0,0xfe040,0xfe03f,0x101fbf,0xfdfc0,0xfdfbf,0x10fe0bf,0x11020bf,0x11020c0,0x10fe0c0,0x2fe041,0x302041,0x301fc1,0x2fdfc1,0x8105fc0,0x8106040,0x810603f,0x8105fbf,0x701fbe,0x70203e,0x6fe03e,0x6fdfbe,0x30fdf3f,0x3101f3f,0x3101f40,0x30fdf40,0x180f9fc0,0x180fa040,0x180fa03f,0x180f9fbf,0x12fe0c1,0x13020c1,0x91060c0,0x91060bf,0x8306041,0x8305fc1,0x3301f41,0x32fdf41,0x182f9fc1,0x182fa041,0x190fa0c0,0x190fa0bf,0x16fe0be,0x17020be,0x870603e,0x8705fbe,0xb105f3f,0xb105f40,0x3701f3e,0x36fdf3e,0x186f9fbe,0x186fa03e,0x1b0f9f3f,0x1b0f9f40,0x93060c1,0x192fa0c1,0x97060be,0xb305f41,0x1b2f9f41,0x196fa0be,0xb705f3e,0x1b6f9f3e,
     11,0x101fc0,0xfe040,0xfdfc0,0x10203f,0x101fbf,0xfe03f,0xfdfbf,0xfdfc1,0x101fc1,0x102041,0xfe041,0x10fe0c0,0x11020c0,0x8106040,0x8105fc0,0x8105fbf,0x810603f,0x11020bf,0x10fe0bf,0x180fa040,0x180f9fc0,0x30fdf40,0x3101f40,0x3101f3f,0x30fdf3f,0x180f9fbf,0x180fa03f,0x6fe03e,0x70203e,0x701fbe,0x6fdfbe,0x8105fc1,0x8106041,0x11020c1,0x10fe0c1,0x180fa041,0x180f9fc1,0x30fdf41,0x3101f41,0x91060c0,0x91060bf,0x190fa0c0,0x190fa0bf,0xb105f40,0xb105f3f,0x8705fbe,0x870603e,0x97020be,0x16fe0be,0x1b0f9f40,0x1b0f9f3f,0x36fdf3e,0xb701f3e,0x1b6f9fbe,0x196fa03e,0x93060c1,0xb305f41,0x192fa0c1,0x1b2f9f41,0x1b2fdfc2,0xb301fc2,0x9302042,0x192fe042,
@@ -77,3 +79,277 @@ const unsigned int voro_base::wl[wl_seq_length*wl_hgridcu]={
     17,0x10203f,0x101fc0,0xfe040,0xfe03f,0x101fbf,0x106040,0x1020c0,0xfdfc0,0x1020bf,0x10603f,0x102041,0xfdfbf,0x105fc0,0xfe0c0,0xfe0bf,0x105fbf,0x1060c0,0x101fc1,0xfe041,0x1060bf,0x106041,0x1020c1,0x2fdfc1,0x305fc1,0x2fe0c1,0x3060c1,0x70203e,0x6fe03e,0x701fbe,0x70603e,0x7020be,0x6fdfbe,0x3101f40,0x3701f3f,0x180fa040,0x186fa03f,0x6fe0be,0x705fbe,0x30fdf40,0x1b0f9fc0,0x186f9fbf,0x1b6fdf3f,0x3105f40,0x3705f3f,0x97060be,0x830a040,0x870a03f,0x1302140,0x180fa0c0,0x186fa0bf,0x170213f,0x192fe140,0x192fa041,0x3301f41,0xb309fc0,0xb709fbf,0x850a0c0,0x9506140,0x196fe13f,0x182f9fc1,0x1b2fdf41,0xb305f41,0x12302042,
     16,0x10203f,0x101fc0,0xfe040,0x102041,0x1020c0,0x106040,0x101fbf,0xfe03f,0xfdfc0,0x101fc1,0x105fc0,0x10603f,0x1020bf,0xfe0c0,0xfe041,0xfdfbf,0x1020c1,0x106041,0x1060c0,0x105fbf,0xfe0bf,0xfdfc1,0x105fc1,0x1060bf,0xfe0c1,0x83060c1,0x70203e,0x3101f40,0x180fa040,0x180fa03f,0x186fe03e,0x701fbe,0x3701f3f,0x30fdf40,0x1b0f9fc0,0x180fa041,0x180fa0c0,0x7020be,0x70603e,0x3105f40,0xb101f41,0x9302042,0x1102140,0x930a040,0x6fdfbe,0x36fdf3f,0x1b6f9fbf,0x190fa0bf,0x196fe0be,0x170213f,0x196fe140,0x182f9fc1,0x1b2fdf41,0xb301fc2,0x1a2fe042,0x192fa0c1,0x8705fbe,0xb705f3f,0xb309fc0,0xa70a03f,0x97060be,0x9706140,0x11302141
 };
+
+
+
+    /** This function is called during container construction. The routine scans
+     * all of the worklists in the wl[] array. For a given worklist of blocks
+     * labeled \f$w_1\f$ to \f$w_n\f$, it computes a sequence \f$r_0\f$ to
+     * \f$r_n\f$ so that $r_i$ is the minimum distance to all the blocks
+     * \f$w_{j}\f$ where \f$j>i\f$ and all blocks outside the worklist. The values
+     * of \f$r_n\f$ is calculated first, as the minimum distance to any block in
+     * the shell surrounding the worklist. The \f$r_i\f$ are then computed in
+     * reverse order by considering the distance to \f$w_{i+1}\f$. */
+    voro_base::voro_base(int nx_,int ny_,int nz_,double boxx_,double boxy_,double boxz_) :
+        nx(nx_), ny(ny_), nz(nz_), nxy(nx_*ny_), nxyz(nxy*nz_), boxx(boxx_), boxy(boxy_), boxz(boxz_),
+        xsp(1/boxx_), ysp(1/boxy_), zsp(1/boxz_), mrad(new double[wl_hgridcu*wl_seq_length]) {
+        const unsigned int b1=1<<21,b2=1<<22,b3=1<<24,b4=1<<25,b5=1<<27,b6=1<<28;
+        const double xstep=boxx/wl_fgrid,ystep=boxy/wl_fgrid,zstep=boxz/wl_fgrid;
+        int i,j,k,lx,ly,lz,q;
+        unsigned int f,*e=const_cast<unsigned int*> (wl);
+        double xlo,ylo,zlo,xhi,yhi,zhi,minr,*radp=mrad;
+        for(zlo=0,zhi=zstep,lz=0;lz<wl_hgrid;zlo=zhi,zhi+=zstep,lz++) {
+            for(ylo=0,yhi=ystep,ly=0;ly<wl_hgrid;ylo=yhi,yhi+=ystep,ly++) {
+                for(xlo=0,xhi=xstep,lx=0;lx<wl_hgrid;xlo=xhi,xhi+=xstep,lx++) {
+                    minr=large_number;
+                    for(q=e[0]+1;q<wl_seq_length;q++) {
+                        f=e[q];
+                        i=(f&127)-64;
+                        j=(f>>7&127)-64;
+                        k=(f>>14&127)-64;
+                        if((f&b2)==b2) {
+                            compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i-1,j,k);
+                            if((f&b1)==0) compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i+1,j,k);
+                        } else if((f&b1)==b1) compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i+1,j,k);
+                        if((f&b4)==b4) {
+                            compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i,j-1,k);
+                            if((f&b3)==0) compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i,j+1,k);
+                        } else if((f&b3)==b3) compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i,j+1,k);
+                        if((f&b6)==b6) {
+                            compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i,j,k-1);
+                            if((f&b5)==0) compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i,j,k+1);
+                        } else if((f&b5)==b5) compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i,j,k+1);
+                    }
+                    q--;
+                    while(q>0) {
+                        radp[q]=minr;
+                        f=e[q];
+                        i=(f&127)-64;
+                        j=(f>>7&127)-64;
+                        k=(f>>14&127)-64;
+                        compute_minimum(minr,xlo,xhi,ylo,yhi,zlo,zhi,i,j,k);
+                        q--;
+                    }
+                    *radp=minr;
+                    e+=wl_seq_length;
+                    radp+=wl_seq_length;
+                }
+            }
+        }
+    }
+
+    /** Computes the minimum distance from a subregion to a given block. If this distance
+     * is smaller than the value of minr, then it passes
+     * \param[in,out] minr a pointer to the current minimum distance. If the distance
+     *                     computed in this function is smaller, then this distance is
+     *                     set to the new one.
+     * \param[out] (xlo,ylo,zlo) the lower coordinates of the subregion being
+     *                           considered.
+     * \param[out] (xhi,yhi,zhi) the upper coordinates of the subregion being
+     *                           considered.
+     * \param[in] (ti,tj,tk) the coordinates of the block. */
+    void voro_base::compute_minimum(double &minr,double &xlo,double &xhi,double &ylo,double &yhi,double &zlo,double &zhi,int ti,int tj,int tk) {
+        double radsq,temp;
+        if(ti>0) {temp=boxx*ti-xhi;radsq=temp*temp;}
+        else if(ti<0) {temp=xlo-boxx*(1+ti);radsq=temp*temp;}
+        else radsq=0;
+
+        if(tj>0) {temp=boxy*tj-yhi;radsq+=temp*temp;}
+        else if(tj<0) {temp=ylo-boxy*(1+tj);radsq+=temp*temp;}
+
+        if(tk>0) {temp=boxz*tk-zhi;radsq+=temp*temp;}
+        else if(tk<0) {temp=zlo-boxz*(1+tk);radsq+=temp*temp;}
+
+        if(radsq<minr) minr=radsq;
+    }
+
+    /** Checks to see whether "%n" appears in a format sequence to determine
+     * whether neighbor information is required or not.
+     * \param[in] format the format string to check.
+     * \return True if a "%n" is found, false otherwise. */
+    bool voro_base::contains_neighbor(const char *format) {
+        char *fmp=(const_cast<char*>(format));
+
+        // Check to see if "%n" appears in the format sequence
+        while(*fmp!=0) {
+            if(*fmp=='%') {
+                fmp++;
+                if(*fmp=='n') return true;
+                else if(*fmp==0) return false;
+            }
+            fmp++;
+        }
+
+        return false;
+    }
+
+
+
+
+
+/** The class constructor sets up the geometry of container, initializing the
+ * minimum and maximum coordinates in each direction, and setting whether each
+ * direction is periodic or not. It divides the container into a rectangular
+ * grid of blocks, and allocates memory for each of these for storing particle
+ * positions and IDs.
+ * \param[in] (ax_,bx_) the minimum and maximum x coordinates.
+ * \param[in] (ay_,by_) the minimum and maximum y coordinates.
+ * \param[in] (az_,bz_) the minimum and maximum z coordinates.
+ * \param[in] (nx_,ny_,nz_) the number of grid blocks in each of the three
+ *			    coordinate directions.
+ * \param[in] (xperiodic_,yperiodic_,zperiodic_) flags setting whether the
+ *                                               container is periodic in each
+ *                                               coordinate direction.
+ * \param[in] init_mem the initial memory allocation for each block.
+ * \param[in] ps_ the number of floating point entries to store for each
+ *                particle. */
+container_base::container_base(double ax_,double bx_,double ay_,double by_,double az_,double bz_,
+        int nx_,int ny_,int nz_,bool xperiodic_,bool yperiodic_,bool zperiodic_,int init_mem,int ps_)
+    : voro_base(nx_,ny_,nz_,(bx_-ax_)/nx_,(by_-ay_)/ny_,(bz_-az_)/nz_),
+    ax(ax_), bx(bx_), ay(ay_), by(by_), az(az_), bz(bz_),
+    xperiodic(xperiodic_), yperiodic(yperiodic_), zperiodic(zperiodic_),
+    id(new int*[nxyz]), p(new double*[nxyz]), co(new int[nxyz]), mem(new int[nxyz]), ps(ps_) {
+    int l;
+    for(l=0;l<nxyz;l++) co[l]=0;
+    for(l=0;l<nxyz;l++) mem[l]=init_mem;
+    for(l=0;l<nxyz;l++) id[l]=new int[init_mem];
+    for(l=0;l<nxyz;l++) p[l]=new double[ps*init_mem];
+}
+
+/** The container destructor frees the dynamically allocated memory. */
+container_base::~container_base() {
+    int l;
+    for(l=0;l<nxyz;l++) delete [] p[l];
+    for(l=0;l<nxyz;l++) delete [] id[l];
+    delete [] id;
+    delete [] p;
+    delete [] co;
+    delete [] mem;
+}
+
+/** The class constructor sets up the geometry of container.
+ * \param[in] (ax_,bx_) the minimum and maximum x coordinates.
+ * \param[in] (ay_,by_) the minimum and maximum y coordinates.
+ * \param[in] (az_,bz_) the minimum and maximum z coordinates.
+ * \param[in] (nx_,ny_,nz_) the number of grid blocks in each of the three
+ *                       coordinate directions.
+ * \param[in] (xperiodic_,yperiodic_,zperiodic_) flags setting whether the
+ *                                               container is periodic in each
+ *                                               coordinate direction.
+ * \param[in] init_mem the initial memory allocation for each block. */
+container::container(double ax_,double bx_,double ay_,double by_,double az_,double bz_,
+    int nx_,int ny_,int nz_,bool xperiodic_,bool yperiodic_,bool zperiodic_,int init_mem)
+    : container_base(ax_,bx_,ay_,by_,az_,bz_,nx_,ny_,nz_,xperiodic_,yperiodic_,zperiodic_,init_mem,3) {}
+
+/** Put a particle into the correct region of the container.
+ * \param[in] n the numerical ID of the inserted particle.
+ * \param[in] (x,y,z) the position vector of the inserted particle. */
+void container::put(int n,double x,double y,double z) {
+    int ijk;
+    if(put_locate_block(ijk,x,y,z)) {
+        id[ijk][co[ijk]]=n;
+        double *pp=p[ijk]+3*co[ijk]++;
+        *(pp++)=x;*(pp++)=y;*pp=z;
+    }
+}
+
+/** This routine takes a particle position vector, tries to remap it into the
+ * primary domain. If successful, it computes the region into which it can be
+ * stored and checks that there is enough memory within this region to store
+ * it.
+ * \param[out] ijk the region index.
+ * \param[in,out] (x,y,z) the particle position, remapped into the primary
+ *                        domain if necessary.
+ * \return True if the particle can be successfully placed into the container,
+ * false otherwise. */
+inline bool container_base::put_locate_block(int &ijk,double &x,double &y,double &z) {
+    if(put_remap(ijk,x,y,z)) {
+        if(co[ijk]==mem[ijk]) add_particle_memory(ijk);
+        return true;
+    }
+    return false;
+}
+
+/** Takes a particle position vector and computes the region index into which
+ * it should be stored. If the container is periodic, then the routine also
+ * maps the particle position to ensure it is in the primary domain. If the
+ * container is not periodic, the routine bails out.
+ * \param[out] ijk the region index.
+ * \param[in,out] (x,y,z) the particle position, remapped into the primary
+ *                        domain if necessary.
+ * \return True if the particle can be successfully placed into the container,
+ * false otherwise. */
+inline bool container_base::put_remap(int &ijk,double &x,double &y,double &z) {
+    int l;
+
+    ijk=step_int((x-ax)*xsp);
+    if(xperiodic) {l=step_mod(ijk,nx);x+=boxx*(l-ijk);ijk=l;}
+    else if(ijk<0||ijk>=nx) return false;
+
+    int j=step_int((y-ay)*ysp);
+    if(yperiodic) {l=step_mod(j,ny);y+=boxy*(l-j);j=l;}
+    else if(j<0||j>=ny) return false;
+
+    int k=step_int((z-az)*zsp);
+    if(zperiodic) {l=step_mod(k,nz);z+=boxz*(l-k);k=l;}
+    else if(k<0||k>=nz) return false;
+
+    ijk+=nx*j+nxy*k;
+    return true;
+}
+
+/** Takes a position vector and attempts to remap it into the primary domain.
+ * \param[out] (ai,aj,ak) the periodic image displacement that the vector is in,
+ *                       with (0,0,0) corresponding to the primary domain.
+ * \param[out] (ci,cj,ck) the index of the block that the position vector is
+ *                        within, once it has been remapped.
+ * \param[in,out] (x,y,z) the position vector to consider, which is remapped
+ *                        into the primary domain during the routine.
+ * \param[out] ijk the block index that the vector is within.
+ * \return True if the particle is within the container or can be remapped into
+ * it, false if it lies outside of the container bounds. */
+inline bool container_base::remap(int &ai,int &aj,int &ak,int &ci,int &cj,int &ck,double &x,double &y,double &z,int &ijk) {
+    ci=step_int((x-ax)*xsp);
+    if(ci<0||ci>=nx) {
+        if(xperiodic) {ai=step_div(ci,nx);x-=ai*(bx-ax);ci-=ai*nx;}
+        else return false;
+    } else ai=0;
+
+    cj=step_int((y-ay)*ysp);
+    if(cj<0||cj>=ny) {
+        if(yperiodic) {aj=step_div(cj,ny);y-=aj*(by-ay);cj-=aj*ny;}
+        else return false;
+    } else aj=0;
+
+    ck=step_int((z-az)*zsp);
+    if(ck<0||ck>=nz) {
+        if(zperiodic) {ak=step_div(ck,nz);z-=ak*(bz-az);ck-=ak*nz;}
+        else return false;
+    } else ak=0;
+
+    ijk=ci+nx*cj+nxy*ck;
+    return true;
+}
+
+/** Increase memory for a particular region.
+ * \param[in] i the index of the region to reallocate. */
+void container_base::add_particle_memory(int i) {
+    int l,nmem=mem[i]<<1;
+
+    // Carry out a check on the memory allocation size
+    if(nmem>max_particle_memory)
+        voro_fatal_error("Absolute maximum memory allocation exceeded",VOROPP_MEMORY_ERROR);
+
+    // Allocate new memory and copy in the contents of the old arrays
+    int *idp=new int[nmem];
+    for(l=0;l<co[i];l++) idp[l]=id[i][l];
+    double *pp=new double[ps*nmem];
+    for(l=0;l<ps*co[i];l++) pp[l]=p[i][l];
+
+    // Update pointers and delete old arrays
+    mem[i]=nmem;
+    delete [] id[i];id[i]=idp;
+    delete [] p[i];p[i]=pp;
+}
+
+}
